requirements.txt
joblib==1.1.0
networkx==2.8.7
numpy==1.23.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
patsy==0.5.2
pgmpy==0.1.20
pyparsing==3.0.9
pyro-api==0.1.2
pyro-ppl==1.8.2
python-dateutil==2.8.2
pytz==2022.4
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
threadpoolctl==3.1.0
torch==1.12.0
tqdm==4.64.0
typing-extensions==4.3.0
